break down onboard file into modular files
