+++ jarvis-app/src/components/onboard/views/IdentifyLevel.js
@@ -1,0 +1,21 @@
+import { useContext } from "react";
+import { WidgetContext } from "../../../modules/widgetContext";
+
+export function IdentifyLevel() {
+    const { userState, setUserState } = useContext(WidgetContext);
+    return (
+      <div>
+        <button
+          onClick={(e) => {
+            e.stopPropagation();
+            setUserState({
+              ...userState,
+              onboardingProgress: userState.onboardingProgress + 1,
+            });
+          }}
+        >
+          b
+        </button>
+      </div>
+    );
+  }
