+++ jarvis-app/src/components/onboard/views/ReasonOfFocus.js
@@ -1,0 +1,21 @@
+import { useContext } from "react";
+import { WidgetContext } from "../../../modules/widgetContext";
+
+export function ReasonOfFocus() {
+    const { userState, setUserState } = useContext(WidgetContext);
+    return (
+      <div>
+        <button
+          onClick={(e) => {
+            e.stopPropagation();
+            setUserState({
+              ...userState,
+              onboardingProgress: userState.onboardingProgress + 1,
+            });
+          }}
+        >
+          f
+        </button>
+      </div>
+    );
+  }
